export ssh public keys under ssh-pub-keys in topo json
