+++ utils/ssh.go
@@ -1,10 +1,13 @@
 package utils
 
 import (
+	"bytes"
 	"os/exec"
 	"regexp"
+	"strings"
 
 	log "github.com/sirupsen/logrus"
+	"golang.org/x/crypto/ssh"
 )
 
 // GetSSHVersion returns the version of the ssh client
@@ -31,4 +34,35 @@ func parseSSHVersion(in string) string {
 	}
 
 	return match[1]
+}
+
+// MarshalSSHPubKeys marshales the ssh public keys
+// and a string slice that contains string representations of the keys.
+func MarshalSSHPubKeys(in []ssh.PublicKey) []string {
+	r := []string{}
+
+	for _, k := range in {
+		// extract the keys in AuthorizedKeys format (e.g. "ssh-rsa <KEY>")
+		ks := bytes.TrimSpace(ssh.MarshalAuthorizedKey(k))
+		r = append(r, string(ks))
+
+	}
+
+	return r
+}
+
+// MarshalAndCatenateSSHPubKeys catenates the ssh public keys
+// and produces a string that can be used in the
+// cli config command to set the ssh public keys
+// for users.
+// Each key value in the catenated string will be double quoted.
+func MarshalAndCatenateSSHPubKeys(in []ssh.PublicKey) string {
+	keysSlice := MarshalSSHPubKeys(in)
+	quotedKeys := make([]string, len(keysSlice))
+
+	for i, k := range keysSlice {
+		quotedKeys[i] = "\"" + k + "\""
+	}
+
+	return strings.Join(quotedKeys, " ")
 }
